ENH: Create PckgDependencies class
PckgDependencies is a dict-like object: it groups together required packages, with package names as keys and related requirements as values.

The class serves two purposes: (1) to define/store dependency requirements for cloudized functions, and (2) to make it easy to write / read / change lists of package requirements inside Python code in a Python-native style.